+++ scripts/add_brownfield.py
@@ -24,7 +24,8 @@ from scripts.add_electricity import flatten, sanitize_carriers
 from scripts.prepare_network import set_transmission_limit
 from scripts.add_existing_baseyear import add_build_year_to_new_assets
 
-from scripts.walloon_scripts.nuclear_helper import add_BEWAL_nuclear
+from scripts.walloon_scripts.nuclear_helper import add_BEWAL_nuclear, retrofit_retired_nuclear
+from scripts.walloon_scripts.BEWAL_potentials import update_BEWAL_potentials
 
 logger = logging.getLogger(__name__)
 idx = pd.IndexSlice
@@ -63,6 +64,7 @@ def add_brownfield(
     dc_i = n.links[n.links.carrier == "DC"].index
     n.links.loc[dc_i, "p_nom_min"] = n_p.links.loc[dc_i, "p_nom_opt"]
 
+    decomissioned_assets = {"Link": None, "Generator": None, "Store": None}
     for c in n_p.iterate_components(["Link", "Generator", "Store"]):
         attr = "e" if c.name == "Store" else "p"
 
@@ -71,7 +73,9 @@ def add_brownfield(
         n_p.remove(c.name, c.df.index[c.df.lifetime == np.inf])
 
         # remove assets whose build_year + lifetime <= year
-        n_p.remove(c.name, c.df.index[c.df.build_year + c.df.lifetime <= year])
+        decomissioned_assets_index = c.df.index[c.df.build_year + c.df.lifetime <= year]
+        decomissioned_assets[c.name] = c.df.loc[decomissioned_assets_index]
+        n_p.remove(c.name, decomissioned_assets_index)
 
         # remove assets if their optimized nominal capacity is lower than a threshold
         # since CHP heat Link is proportional to CHP electric Link, make sure threshold is compatible
@@ -158,6 +162,8 @@ def add_brownfield(
             )
             n.links.loc[gas_pipes_i, "p_nom"] = remaining_capacity
             n.links.loc[gas_pipes_i, "p_nom_max"] = remaining_capacity
+
+    return decomissioned_assets
 
 
 def disable_grid_expansion_if_limit_hit(n):
@@ -365,7 +371,7 @@ if __name__ == "__main__":
     if snakemake.params.tes and snakemake.params.dynamic_ptes_capacity:
         update_dynamic_ptes_capacity(n, n_p, year)
 
-    add_brownfield(
+    decomissioned_assets = add_brownfield(
         n,
         n_p,
         year,
@@ -385,6 +391,22 @@ if __name__ == "__main__":
         n=n,
         walloon_nuclear_config=snakemake.config["electricity"]["extendable_carriers"].get("Walloon", {}),
         planning_horizon=planning_horizon,
+        extendable_nuclear_nodes=(
+            snakemake.config["electricity"]["extendable_carriers"]
+            .get("extendable_nuclear_links", {})
+        ),
+    )
+
+    decomissioned_nuclear = (
+        decomissioned_assets["Link"].query("carrier == 'nuclear'")
+    )
+
+    retrofit_retired_nuclear(
+        n,
+        decomissioned_nuclear,
+        int(snakemake.wildcards.planning_horizons),
+        costs = load_costs(snakemake.input.costs),
+        MILP = True,
     )
 
     kind = snakemake.params.transmission_limit[planning_horizon][0]